Move stepper into a separate component
