+++ apps/web/components/real.ts
@@ -71,3 +71,4 @@ export * from './SimpleSlider/SimpleSlider'
 export * from './Image'
 export * from './Webreader/Webreader'
 export * from './AppendedArticleComponents/AppendedArticleComponents'
+export * from './Stepper/Stepper'
